+++ mobile/src/screens/Feed/FeedNav.tsx
@@ -1,15 +1,15 @@
-// mobile/src/screens/LandingScreen.tsx
+// mobile/src/screens/Feed/FeedNav.tsx
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { StackParams } from '../types/StackParams';
-import { colors, typography } from '../styles';
-import IssueDetailScreen from './IssueDetailScreen';
-import ErrorScreen from './ErrorScreen';
-import LandingScreen from './LandingScreen';
+import { StackParams } from '../../types/StackParams';
+import { colors, typography } from '../../styles';
+import FeedScreen from './FeedScreen';
+import LeaderBoardScreen from './LeaderboardScreen';
+import IssueDetailScreen from '../Misc/IssueDetailScreen';
+import ErrorScreen from '../Misc/ErrorScreen';
 
 const Stack = createNativeStackNavigator<StackParams>();
 
-export default function LandingScreenNav() {
-
+export default function FeedNav() {
     return (
         <Stack.Navigator screenOptions={{
             headerStyle: {
@@ -23,7 +23,12 @@ export default function LandingScreenNav() {
 
         }}
         >
-            <Stack.Screen name="Nearby Issues" component={LandingScreen} />
+            <Stack.Screen name="Feed" component={FeedScreen}
+                options={{
+                    headerShown: true
+                }} />
+            <Stack.Screen name="Leaderboard" component={LeaderBoardScreen} />
+
             <Stack.Screen name="Issue Details" component={IssueDetailScreen} />
             <Stack.Screen name="Error" component={ErrorScreen} />
         </Stack.Navigator>
