+++ mobile/src/screens/Landing/LandingScreenNav.tsx
@@ -1,0 +1,32 @@
+// mobile/src/screens/Landing/LandingScreen.tsx
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import { StackParams } from '../../types/StackParams';
+import { colors, typography } from '../../styles';
+import IssueDetailScreen from '../Misc/IssueDetailScreen';
+import ErrorScreen from '../Misc/ErrorScreen';
+import LandingScreen from './LandingScreen';
+
+const Stack = createNativeStackNavigator<StackParams>();
+
+export default function LandingScreenNav() {
+
+    return (
+        <Stack.Navigator screenOptions={{
+            headerStyle: {
+                backgroundColor: colors.background,
+            },
+            headerTintColor: colors.textPrimary,
+            headerTitleStyle: {
+                fontWeight: typography.weightRegular,
+            },
+            headerTitleAlign: "left"
+
+        }}
+        >
+            <Stack.Screen name="Nearby Issues" component={LandingScreen} />
+            <Stack.Screen name="Issue Details" component={IssueDetailScreen} />
+            <Stack.Screen name="Error" component={ErrorScreen} />
+        </Stack.Navigator>
+    );
+
+}
